Move cli utility funcitons to utils.py and remove __future__ import annotations
